+++ doc/whatsnew.rst
@@ -45,6 +45,9 @@ v0.5 (Jun 2023)
 
 - Fix :func:`pyriemann.datasets.sample_gaussian_spd` used with ``sampling_method=rejection`` on 2D matrices. :pr:`250` by :user:`mhurte`
 
+- Add :func:`pyriemann.utils.test.is_real_type` to check the type of input arrays and
+  add :func:`pyriemann.utils.covariance.covariance_scm` allowing to process complex-valued inputs for 'scm' covariance estimator. :pr:`251` by :user:`qbarthelemy`
+
 v0.4 (Feb 2023)
 ---------------